// api/src/modules/order/domain/enums/order-status.enum.ts
export enum OrderStatus {
  DRAFT = 'draft',
  OPEN = 'open',
  IN_KITCHEN = 'in_kitchen',
  PARTIALLY_SERVED = 'partially_served',
  SERVED = 'served',
  BILL_REQUESTED = 'bill_requested',
  PAYMENT_IN_PROGRESS = 'payment_in_progress',
  CLOSED = 'closed',
  CANCELLED = 'cancelled',
  VOIDED = 'voided',
}